backend/pyproject.toml
# pyproject.toml
[project]
name = "mlproj"
version = "0.1.0"
description = "Gym Workout Recommender – FastAPI + fastapi-users + SQLModel"
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.18",
    "fastapi-users[sqlalchemy2]>=12.1.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "python-multipart>=0.0.9",
    "aiosqlite>=0.19.0",
    "python-dotenv>=1.2.1",
    "asyncpg>=0.27.0"
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "pytest>=7.0.0"
]
